backend/requirements.txt
# ================================================
# MAI-PAEP Backend Requirements
# ================================================

# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1

# Redis (Caching)
redis[hiredis]==5.0.1

# AI APIs
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2
groq==0.4.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# HTTP Client
aiohttp==3.9.1
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pydantic-core==2.14.6

# Monitoring and Logging
python-json-logger==2.0.7

# ML/NLP (copied from ml/requirements.txt for convenience)
torch==2.1.2
transformers==4.37.0
sentence-transformers==2.2.2
numpy==1.26.3
scipy==1.11.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
